Sort lines by time as an integer (#78)

+++ web/models/stream_uploader.ex
@@ -26,7 +26,7 @@ defmodule Streamr.StreamUploader do
       from stream_data
       left join lateral unnest(lines) as line on true
       where stream_id = #{stream.id}
-      order by line->>'time' asc
+      order by (line->>'time')::int asc
     """
   end
